+++ Services/Repositories/Implimentations/DanhMuc/DoiTuongService.cs
@@ -130,7 +130,10 @@ namespace Services.Repositories.Implimentations.DanhMuc
                 string filePath = Path.Combine(destPath, fileName);
                 package.SaveAs(new FileInfo(filePath));
                 byte[] fileByte = File.ReadAllBytes(filePath);
-                File.Delete(filePath);
+                if (File.Exists(filePath))
+                {
+                    File.Delete(filePath);
+                }
 
                 return new FileReturn
                 {
@@ -193,7 +196,7 @@ namespace Services.Repositories.Implimentations.DanhMuc
         {
             var query = new List<DoiTuongViewModel>();
 
-            query = _mp.Map<List<DoiTuongViewModel>>(await _db.DoiTuongs.AsNoTracking().Where(x => x.IsNhanVien == true && x.Status==true).ToListAsync());
+            query = _mp.Map<List<DoiTuongViewModel>>(await _db.DoiTuongs.AsNoTracking().Where(x => x.IsNhanVien == true && x.Status == true).ToListAsync());
 
             return query;
         }
